Use ManageLabels activity as StudyPad listing

+++ app/src/main/java/net/bible/android/view/activity/bookmark/ManageLabels.kt
@@ -19,22 +19,35 @@ package net.bible.android.view.activity.bookmark
 
 import android.annotation.SuppressLint
 import android.app.Activity
+import android.app.AlertDialog
 import android.content.Intent
 import android.graphics.Color
 import android.os.Bundle
 import android.util.Log
+import android.view.Menu
+import android.view.MenuItem
 import android.view.View
 import android.widget.ListView
 import kotlinx.android.synthetic.main.manage_labels.*
+import kotlinx.android.synthetic.main.manage_labels_list_item.*
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import net.bible.android.activity.R
 import net.bible.android.control.bookmark.BookmarkControl
 import net.bible.android.control.event.ABEventBus
 import net.bible.android.control.event.ToastEvent
+import net.bible.android.control.page.window.ActiveWindowPageManagerProvider
 import net.bible.android.database.bookmarks.BookmarkEntities
+import net.bible.android.view.activity.base.Dialogs
 import net.bible.android.view.activity.base.ListActivityBase
 import net.bible.service.common.CommonUtils
+import net.bible.service.download.FakeBookFactory
+import net.bible.service.sword.StudyPadKey
 import java.util.*
 import javax.inject.Inject
+import kotlin.coroutines.resume
+import kotlin.coroutines.suspendCoroutine
 import kotlin.random.Random.Default.nextInt
 
 /**
@@ -46,11 +59,13 @@ class ManageLabels : ListActivityBase() {
     private val labels: MutableList<BookmarkEntities.Label> = ArrayList()
     @Inject lateinit var bookmarkControl: BookmarkControl
     @Inject lateinit var labelDialogs: LabelDialogs
+    @Inject lateinit var activeWindowPageManagerProvider: ActiveWindowPageManagerProvider
 
     var showUnassigned = false
     var showCheckboxes = false
     var selectMultiple = false
     var hasResetButton = false
+    var studyPadMode = false
     private val deletedLabels = mutableSetOf<Long>()
 
     @SuppressLint("MissingSuperCall")
@@ -64,20 +79,63 @@ class ManageLabels : ListActivityBase() {
             checkedLabels.addAll(selectedLabelIds.toList())
         }
 
+        studyPadMode = intent.getBooleanExtra("studyPadMode", false)
         hasResetButton = intent.getBooleanExtra("resetButton", false)
         resetButton.visibility = if(hasResetButton) View.VISIBLE else View.GONE
         showUnassigned = intent.getBooleanExtra("showUnassigned", false)
         val title = intent.getStringExtra("title")
         selectMultiple = checkedLabels.size > 1 || CommonUtils.sharedPreferences.getBoolean("assignLabelsSelectMultiple", false)
         selectMultipleSwitch.isChecked = selectMultiple
+        selectMultipleSwitch.visibility = if(showCheckboxes) View.VISIBLE else View.GONE
         selectMultipleSwitch.setOnCheckedChangeListener { _, isChecked ->
             selectMultiple = isChecked
             CommonUtils.sharedPreferences.edit().putBoolean("assignLabelsSelectMultiple", selectMultiple).apply()
         }
+
+        if(studyPadMode) {
+            setTitle(getString(R.string.studypads))
+        }
+
         if(title!=null) setTitle(title)
+
         listView.choiceMode = ListView.CHOICE_MODE_MULTIPLE
         loadLabelList()
         listAdapter = ManageLabelItemAdapter(this, LIST_ITEM_TYPE, labels, this, checkedLabels, showCheckboxes)
+    }
+
+    override fun onPrepareOptionsMenu(menu: Menu): Boolean {
+        menuInflater.inflate(R.menu.journals_options_menu, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        var isHandled = false
+        when(item.itemId){
+            R.id.help -> {
+                CommonUtils.showHelp(this, listOf(R.string.help_studypads_title))
+            }
+        }
+        if (!isHandled) {
+            isHandled = super.onOptionsItemSelected(item)
+        }
+        return isHandled
+    }
+
+    private fun studyPadSelected(journal: BookmarkEntities.Label) {
+        Log.d(TAG, "Journal selected:" + journal.name)
+        try {
+            activeWindowPageManagerProvider.activeWindowPageManager.setCurrentDocumentAndKey(FakeBookFactory.journalDocument, StudyPadKey(journal))
+        } catch (e: Exception) {
+            Log.e(TAG, "Error on attempt to show journal", e)
+            Dialogs.instance.showErrorMsg(R.string.error_occurred, e)
+        }
+    }
+
+    override fun onListItemClick(l: ListView, v: View, position: Int, id: Long) {
+        if(studyPadMode) {
+            okay(labels[position])
+        }
+        super.onListItemClick(l, v, position, id)
     }
 
     private val checkedLabels = mutableSetOf<Long>()
@@ -118,13 +176,34 @@ class ManageLabels : ListActivityBase() {
     }
 
     fun onOkay(v: View?) {
+        okay()
+    }
+
+    private fun okay(selected: BookmarkEntities.Label? = null) = GlobalScope.launch(Dispatchers.Main) {
         Log.i(TAG, "Okay clicked")
+
+        if(deletedLabels.size > 0 && !askConfirmation()) return@launch
         val result = Intent()
         bookmarkControl.deleteLabels(deletedLabels.toList())
         val labelIds = checkedLabels.toLongArray()
         result.putExtra(BookmarkControl.LABEL_IDS_EXTRA, labelIds)
         setResult(Activity.RESULT_OK, result)
+        if(selected != null) {
+            studyPadSelected(selected)
+        }
         finish()
+    }
+
+    private suspend fun askConfirmation()  = suspendCoroutine<Boolean> {
+        AlertDialog.Builder(this)
+            .setMessage(getString(R.string.confirm_delete_study_pads, deletedLabels.size))
+            .setPositiveButton(R.string.yes) { _, _ ->
+                it.resume(true)
+            }
+            .setCancelable(true)
+            .setOnCancelListener { _ -> it.resume(false) }
+            .setNegativeButton(R.string.cancel) { _, _ -> it.resume(false)}
+            .show()
     }
 
     fun onReset(v: View?) {
